[SC] Align the order of input arguments and update contracts (#2482)
* Add exception code for new method of remote backend

* Updated default gaslimit to deploy

* Aligned the order of arguments

- Aligned the order of arguments
- Added service chain option to avoid to check local balance
- Fixed minor bug int64 -> 256
- update test code for new order of the arguments

* reflected test code

+++ contracts/servicechain_nft/sc_nft.sol
@@ -12,7 +12,7 @@ contract ServiceChainNFT is ERC721Full("ServiceChainNFT", "SCN"), Ownable {
     constructor (address _bridge) public { bridge = _bridge; }
 
     // Owner mints the NFT to the user.
-    function register(address _user, uint64 _tokenId) onlyOwner external {
+    function register(address _user, uint256 _tokenId) onlyOwner external {
         _mint(_user, _tokenId);
     }
